create only one gstreamer player
The gstreamer player should be created only once and then be reused when
more than one file is to be played. This fixes a bug where pausing and
aborting playback of one song would still keep the playbin and resume it
when the next song in the list would be played, paused and resumed.